app/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
anthropic>=0.47.0
openai>=1.59.0
tavily-python>=0.5.0
twilio>=8.59.0
boto3>=1.36.0
loguru>=0.7.0
elevenlabs>=0.3.0
pydub>=0.25.1 # winget install ffmpeg
ffmpeg-downloader # ffdl install --add-path
requests>=2.31.0
beautifulsoup4>=4.12.0
pytz>=2024.1
python-telegram-bot[job-queue]>=20.0
faiss-cpu>=1.7.4
numpy>=1.24.0
docker>=6.1.3
